+++ lib/hyperion/util.rb
@@ -1,0 +1,11 @@
+class Hyperion
+  class Util
+    def self.nil_if_error
+      begin
+        yield
+      rescue StandardError
+        return nil
+      end
+    end
+  end
+end